backend/requirements.txt
alembic==1.13.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.10
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.5.2
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
flasgger==0.9.7.1
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
pkgutil-resolve-name==1.3.10
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.9.2
pydantic-core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
tomli==2.0.2
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
werkzeug==3.0.4
zipp==3.20.2
